+++ RELEASE_NOTES.md
@@ -1,3 +1,30 @@
+# Czatbox TT 0.1.1
+
+## Główne funkcje
+
+- przybornik live został zastąpiony widgetami. Od teraz możesz sprawdzić moderację online na czacie, statystyki i top 5 giftujących osób
+- wszystkie widgety są elastyczne tj. możesz je ukryć, pokazać, przypiąć do pulpitu gdy aplikacja jest pomniejszona
+- poprawione i rozbudowane centrum archiwizacji czatu live, z podziałem na filtry, podsumowaniami ilości zdarzeń, wiadomości na czacie, monetami, moderatorami online podczas sesji
+- archiwum można eksportować do pliku *.txt, kasować, odświeżać.
+- język polski, angielski i niemiecki aplikacji
+
+## Poprawki
+
+- poprawiony został komunikat o wysłanej skrzyneczce gdy pękała na czacie "unknow wysyła skrzyneczkę"
+- optymalizacja działania programu
+
+## Znane błędy
+
+- mnożnik bitewek to funkcja testowa, i działa na tak zwaną trytytkę, dlatego czasem w ostatniej minucie się buguje i pojawia mimo jego braku
+
+## Co dalej
+
+- rozbudowa systemu widgetów
+- dodatkowe języki odczytu czatu TTS
+- wyróżnienie super fanów na czacie
+- optymalizacja połączenia z danym twórcą
+- powiadomienia dźwiękowe dla większych prezentów
+
 # Czatbox TT 0.1.0
 
 Pierwsza publiczna wersja programu Czatbox TT.